R: reorganize a few tests, rename RegisterNugetResolvedPackageReferences -> RegisterResolvedPackageReferences
