+++ Microsoft.DotNet.Interactive.CSharp.Tests/SubmissionParsingTests.cs
@@ -8,7 +8,6 @@ using FluentAssertions;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.DotNet.Interactive.Commands;
-using Microsoft.DotNet.Interactive.Tests;
 using Microsoft.DotNet.Interactive.Tests.Utility;
 using Xunit;
