+++ Microsoft.DotNet.Interactive/ISupportNuget.cs
@@ -12,12 +12,12 @@ namespace Microsoft.DotNet.Interactive
         // Set assemblyProbingPaths, nativeProbingRoots for Kernel.
         // These values are functions that return the list of discovered assemblies, and package roots
         // They are used by the dependecymanager for Assembly and Native dll resolving
-        public abstract void Initialize(AssemblyResolutionProbe assemblyProbingPaths, NativeResolutionProbe nativeProbingRoots);
+        public void Initialize(AssemblyResolutionProbe assemblyProbingPaths, NativeResolutionProbe nativeProbingRoots);
 
-        public abstract void RegisterNugetResolvedPackageReferences(IReadOnlyList<ResolvedPackageReference> packageReferences);
+        public void RegisterResolvedPackageReferences(IReadOnlyList<ResolvedPackageReference> packageReferences);
 
         // Summary:
         //     Resolve reference for a list of package manager lines
-        public abstract IResolveDependenciesResult Resolve(IEnumerable<string> packageManagerTextLines, string executionTfm, ResolvingErrorReport reportError);
+        public IResolveDependenciesResult Resolve(IEnumerable<string> packageManagerTextLines, string executionTfm, ResolvingErrorReport reportError);
     }
 }
